refactor: reorganize parser logic and improve error handling
- Moved parser functions into the parser module for better structure.
- Enhanced error handling in the parsing process to provide clearer feedback.
- Streamlined the parsing logic for JSON values, arrays, and objects.
